# Optimize base58 decoding.