--- a/base58/base58.go
+++ b/base58/base58.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2013-2014 Conformal Systems LLC.
+// Copyright (c) 2013-2015 Conformal Systems LLC.
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -6,11 +6,9 @@ package base58
 
 import (
 	"math/big"
-	"strings"
 )
 
-// alphabet is the modified base58 alphabet used by Bitcoin.
-const alphabet = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz"
+//go:generate go run genalphabet.go
 
 var bigRadix = big.NewInt(58)
 var bigZero = big.NewInt(0)
@@ -20,16 +18,15 @@ func Decode(b string) []byte {
 	answer := big.NewInt(0)
 	j := big.NewInt(1)
 
+	scratch := new(big.Int)
 	for i := len(b) - 1; i >= 0; i-- {
-		tmp := strings.IndexAny(alphabet, string(b[i]))
-		if tmp == -1 {
+		tmp := b58[b[i]]
+		if tmp == 255 {
 			return []byte("")
 		}
-		idx := big.NewInt(int64(tmp))
-		tmp1 := big.NewInt(0)
-		tmp1.Mul(j, idx)
-
-		answer.Add(answer, tmp1)
+		scratch.SetInt64(int64(tmp))
+		scratch.Mul(j, scratch)
+		answer.Add(answer, scratch)
 		j.Mul(j, bigRadix)
 	}
 
@@ -37,7 +34,7 @@ func Decode(b string) []byte {
 
 	var numZeros int
 	for numZeros = 0; numZeros < len(b); numZeros++ {
-		if b[numZeros] != alphabet[0] {
+		if b[numZeros] != alphabetIdx0 {
 			break
 		}
 	}
@@ -65,7 +62,7 @@ func Encode(b []byte) string {
 		if i != 0 {
 			break
 		}
-		answer = append(answer, alphabet[0])
+		answer = append(answer, alphabetIdx0)
 	}
 
 	// reverse