@jrick 2015-01-24
Don't merge this yet. It may panic on non-ASCII input due to indexing outside of the 7-bit character range.  I'll fix by extending the digit table to 256 bytes.