@davecgh 2015-01-24
I was just about to post that it could panic on non-ASCII.