@davecgh 2015-01-24
Also, this needs license headers.