@davecgh 2015-01-24
Might want to update the PR description with the new timing for historical purposes.  Looks good.